requirements.txt
python>=3.9
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=0.15.0
ray[rllib]==2.7.0
stable-baselines3[extra]==2.2.1
gymnasium==0.29.1
networkx>=3.0
osmnx>=1.2.1
fastapi>=0.103.0
uvicorn[standard]>=0.23.2
pydantic>=2.4.2
scipy>=1.11.2
numpy>=1.25.2
pandas>=2.1.1
pyyaml>=6.0.1
kafka-python>=2.0.2
pyspark>=3.4.1
psycopg2-binary>=2.9.9
redis>=5.0.1
requests>=2.31.0
prometheus_client>=0.17.1
aiohttp>=3.8.5
tritonclient[http,grpc]>=2.34.0
matplotlib>=3.8.0
seaborn>=0.13.0
simpy>=4.0.1
graphhopper-client>=0.1.0
folium>=0.14.0
plotly>=5.17.0
streamlit>=1.28.0
jupyter>=1.0.0
ipykernel>=6.25.0
pre-commit>=3.5.0
black>=23.9.1
flake8>=6.1.0
pytest>=7.4.3
mypy>=1.6.1
isort>=5.12.0
git+https://github.com/supremeloki/lono_libs.git
